# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CompanySearchResult": {
        "description": "A single fuzzy-search hit (FR-001).",
        "properties": {
          "cik": {
            "title": "Cik",
            "type": "string"
          },
          "exchange": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exchange"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "cik",
          "name",
          "score"
        ],
        "title": "CompanySearchResult",
        "type": "object"
      },
...
      "Factor": {
        "description": "A bull or bear factor.",
        "properties": {
          "claims": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Claims",
            "type": "array"
          },
          "counterpoint_source_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Counterpoint Source Ids",
            "type": "array"
          },
          "monitoring_indicator": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Monitoring Indicator"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title",
          "claims"
        ],
        "title": "Factor",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobStage": {
        "properties": {
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "running",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "name",
          "status"
        ],
        "title": "JobStage",
        "type": "object"
      },
      "JobState": {
        "properties": {
          "progress": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "integer"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage"
          },
          "stages": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/JobStage"
            },
            "title": "Stages",
            "type": "array"
          },
          "status": {
            "enum": [
              "queued",
              "processing",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "progress"
        ],
        "title": "JobState",
        "type": "object"
      },
...
      "ManagementTopicStatement": {
        "properties": {
          "period": {
            "title": "Period",
            "type": "string"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "speaker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speaker"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "topic": {
            "enum": [
              "ai_demand",
              "revenue_outlook",
              "margins",
              "capital_spending",
              "competition",
              "regulation",
              "product_roadmap"
            ],
            "title": "Topic",
            "type": "string"
          }
        },
        "required": [
          "topic",
          "period",
          "text",
          "source_ids"
        ],
        "title": "ManagementTopicStatement",
        "type": "object"
      },
...
      "NewsItemSummary": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "news_id": {
            "title": "News Id",
            "type": "string"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "news_id",
          "summary",
          "category",
          "source_ids"
        ],
        "title": "NewsItemSummary",
        "type": "object"
      },
...
      "ResearchAccepted": {
        "description": "202 response when a research job is queued.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "research_id": {
            "title": "Research Id",
            "type": "string"
          },
          "status": {
            "default": "processing",
            "enum": [
              "queued",
              "processing",
              "done",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "research_id",
          "job_id"
        ],
        "title": "ResearchAccepted",
        "type": "object"
      },
      "ResearchClaim": {
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "confidence_label": {
            "anyOf": [
              {
                "enum": [
                  "strongly_supported",
                  "supported",
                  "limited_evidence"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Label"
          },
          "internal_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Internal Confidence"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "text",
          "source_ids"
        ],
        "title": "ResearchClaim",
        "type": "object"
      },
      "ResearchReport": {
        "properties": {
          "balance_sheet_analysis": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Balance Sheet Analysis",
            "type": "array"
          },
          "bear_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "title": "Bear Factors",
            "type": "array"
          },
          "bull_factors": {
            "items": {
              "$ref": "#/components/schemas/Factor"
            },
            "title": "Bull Factors",
            "type": "array"
          },
          "cash_flow_analysis": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Cash Flow Analysis",
            "type": "array"
          },
          "company_overview": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Company Overview",
            "type": "array"
          },
          "executive_summary": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Executive Summary",
            "type": "array"
          },
          "insufficient_evidence_sections": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Insufficient Evidence Sections",
            "type": "array"
          },
          "management_commentary": {
            "items": {
              "$ref": "#/components/schemas/ManagementTopicStatement"
            },
            "title": "Management Commentary",
            "type": "array"
          },
          "news_summary": {
            "items": {
              "$ref": "#/components/schemas/NewsItemSummary"
            },
            "title": "News Summary",
            "type": "array"
          },
          "profitability_analysis": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Profitability Analysis",
            "type": "array"
          },
          "revenue_analysis": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Revenue Analysis",
            "type": "array"
          },
          "risks": {
            "items": {
              "$ref": "#/components/schemas/Risk"
            },
            "title": "Risks",
            "type": "array"
          },
          "valuation_analysis": {
            "items": {
              "$ref": "#/components/schemas/ResearchClaim"
            },
            "title": "Valuation Analysis",
            "type": "array"
          }
        },
        "required": [
          "executive_summary",
          "company_overview",
          "revenue_analysis",
          "profitability_analysis",
          "balance_sheet_analysis",
          "cash_flow_analysis",
          "valuation_analysis",
          "news_summary",
          "risks",
          "management_commentary",
          "bull_factors",
          "bear_factors"
        ],
        "title": "ResearchReport",
        "type": "object"
      },
      "ResearchReportEnvelope": {
        "properties": {
          "data_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data Version"
          },
          "generated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Generated At"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "report": {
            "$ref": "#/components/schemas/ResearchReport"
          },
          "research_id": {
            "title": "Research Id",
            "type": "string"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "research_id",
          "ticker",
          "report"
        ],
        "title": "ResearchReportEnvelope",
        "type": "object"
      },
      "ResearchRequest": {
        "properties": {
          "force_refresh": {
            "default": false,
            "title": "Force Refresh",
            "type": "boolean"
          },
          "ticker": {
            "title": "Ticker",
            "type": "string"
          }
        },
        "required": [
          "ticker"
        ],
        "title": "ResearchRequest",
        "type": "object"
      },
      "Risk": {
        "properties": {
          "category": {
            "enum": [
              "business",
              "competitive",
              "regulatory",
              "supply_chain",
              "customer_concentration",
              "geographic",
              "technology",
              "litigation",
              "capital_allocation",
              "macroeconomic"
            ],
            "title": "Category",
            "type": "string"
          },
          "change_status": {
            "anyOf": [
              {
                "enum": [
                  "new",
                  "unchanged",
                  "reworded",
                  "removed"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Change Status"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "evidence_label": {
            "anyOf": [
              {
                "enum": [
                  "strongly_supported",
                  "supported",
                  "limited_evidence"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence Label"
          },
          "source_ids": {
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Source Ids",
            "type": "array"
          }
        },
        "required": [
          "category",
          "description",
          "source_ids"
        ],
        "title": "Risk",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Contract-first API for InvestiLens (spec \u00a724).",
    "title": "InvestiLens API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/companies/search": {
      "get": {
        "description": "Fuzzy company search by name or ticker (FR-001).",
        "operationId": "search_companies_api_v1_companies_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "minLength": 1,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanySearchResult"
                  },
                  "title": "Response Search Companies Api V1 Companies Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Companies",
        "tags": [
          "Company"
        ]
      }
    },
...
    "/api/v1/research": {
      "post": {
        "operationId": "create_research_api_v1_research_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Research",
        "tags": [
          "Research"
        ]
      }
    },
    "/api/v1/research/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_api_v1_research_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "Research"
        ]
      }
    },
    "/api/v1/research/jobs/{job_id}/events": {
      "get": {
        "description": "SSE stream of job progress (API-004). Emits the current state, then closes.",
        "operationId": "job_events_api_v1_research_jobs__job_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Events",
        "tags": [
          "Research"
        ]
      }
    },
    "/api/v1/research/{research_id}": {
      "get": {
        "operationId": "get_research_api_v1_research__research_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "research_id",
            "required": true,
            "schema": {
              "title": "Research Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchReportEnvelope"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Research",
        "tags": [
          "Research"
        ]
      }
    },
    "/api/v1/research/{research_id}/diff": {
      "get": {
        "operationId": "diff_research_api_v1_research__research_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_id",
            "required": true,
            "schema": {
              "title": "Research Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "against",
            "required": true,
            "schema": {
              "title": "Against",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Research",
        "tags": [
          "Research"
        ]
      }
    },
    "/api/v1/research/{research_id}/export": {
      "get": {
        "operationId": "export_research_api_v1_research__research_id__export_get",
        "parameters": [
          {
            "in": "path",
            "name": "research_id",
            "required": true,
            "schema": {
              "title": "Research Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "default": "pdf",
              "title": "Format",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Research",
        "tags": [
          "Research"
        ]
      }
    },
...
  }
}